proxyIndexer updated to hdlrIndexer
proxyIndexer modified to return the actual handler that needs to be passed to the Proxy object. Handler is now represented in a functional/class context so it must be initialized with the new keyword before being used as the handler parameter. The object will also expose the createProxy() method to easily retrieve a proxied object on a data object. The change was made so that the user can determine how best to use the handler, whether that be in it's own proxy, or by passing it to another custom handler that combines multiple handles in a complex operation (like my use-case).

var pobj = new jsonQuery.hdlrIndexer().createProxy(obj);
OR
var pobj = new Proxy(obj, new jsonQuery.hdlrIndexer());
OR
var pobj = new Proxy(obj, new MyCustomHandler(myFirstHndl, new jsonQuery.hdlrIndexer());